格式化代码以便符合google YAPF 和 flake8 规范
