+++ src/Demo.py
@@ -1,7 +1,6 @@
-#coding:utf-8
+# coding:utf-8
 
-
-if __name__=="__main__":
+if __name__ == "__main__":
     name = input("What\'s your name?\n")
-    print("Hello %s"%name)
+    print("Hello %s" % name)
     input("按任意键结束")
